Cleaned Up Recommender
